requirements.txt
django==3.0.5
django-countries==5.5
django-model-utils==4.0.0
lxml==4.5.0
Pillow==7.0.0
stripe==2.44.0
mysqlclient==2.0.1